remove some codes; add some todos
