--- eduagent/api/endpoints.py
@@ -1,438 +1,0 @@
-import uuid
-from datetime import UTC, datetime
-from typing import Any
-
-from fastapi import (
-    APIRouter,
-    HTTPException,
-)
-
-from .schemas import (
-    # Assessment & Feedback
-    AssessmentRequest,
-    AssessmentResponse,
-    AssessmentResult,
-    BatchOperationResponse,
-    # Enums
-    CognitiveLevel,
-    DifficultyControlRequest,
-    DifficultyControlResponse,
-    DifficultyLevel,
-    # Exercise & Practice
-    ExerciseCreateRequest,
-    ExerciseResponse,
-    FeedbackRequest,
-    FeedbackResponse,
-    GeneratedQuestion,
-    # System
-    HealthCheckResponse,
-    # Knowledge Extraction
-    KnowledgeExtractionResponse,
-    KnowledgeGraphResponse,
-    LoginRequest,
-    LoginResponse,
-    MistakeAnalysisRequest,
-    MistakeAnalysisResponse,
-    # Analytics
-    PerformanceAnalyticsRequest,
-    PerformanceAnalyticsResponse,
-    PracticeSessionRequest,
-    PracticeSessionResponse,
-    # Question Generation
-    QuestionGenerationRequest,
-    QuestionGenerationResponse,
-    QuestionType,
-    SubjectArea,
-    # User Management
-    UserCreateRequest,
-    UserResponse,
-    UserRole,
-)
-
-router = APIRouter()
-
-
-# ============ Knowledge Extraction Endpoints ============
-@router.post("/textbook/upload")
-async def upload_textbook() -> KnowledgeExtractionResponse:
-    """
-    Upload textbook for knowledge extraction and multi-modal analysis
-    """
-    extraction_id = str(uuid.uuid4())
-
-    # Background task for processing would be added here
-    return KnowledgeExtractionResponse(
-        extraction_id=extraction_id,
-        status="processing",
-        extracted_concepts=[],
-        created_at=datetime.now(UTC),
-    )
-
-
-@router.get("/knowledge/extraction/{extraction_id}")
-async def get_extraction_status(extraction_id: str) -> KnowledgeExtractionResponse:
-    """
-    Get status of knowledge extraction process
-    """
-    # Mock implementation
-    return KnowledgeExtractionResponse(
-        extraction_id=extraction_id,
-        status="completed",
-        extracted_concepts=[{"concept": "sample", "confidence": 0.95}],
-        created_at=datetime.now(UTC),
-    )
-
-
-@router.get("/knowledge/graph/{textbook_id}")
-async def get_knowledge_graph(textbook_id: str) -> KnowledgeGraphResponse:
-    """
-    Retrieve 3D knowledge graph for a textbook
-    """
-    # Mock implementation
-    _ = textbook_id  # Avoid unused variable warning
-    return KnowledgeGraphResponse(
-        knowledge_points=[], ability_targets=[], common_mistakes=[], graph_structure={}
-    )
-
-
-# ============ Question Generation Endpoints ============
-@router.post("/questions/generate")
-async def generate_questions(
-    request: QuestionGenerationRequest,
-) -> QuestionGenerationResponse:
-    """
-    Generate educational questions based on knowledge points and constraints
-    """
-    # Mock implementation
-    sample_question = GeneratedQuestion(
-        id=str(uuid.uuid4()),
-        question_text="Sample generated question",
-        question_type=request.question_type,
-        difficulty=request.difficulty,
-        cognitive_level=CognitiveLevel.UNDERSTANDING,
-        knowledge_point_ids=request.knowledge_point_ids,
-        estimated_difficulty=0.6,
-        options=None,
-        correct_answer=None,
-        explanation=None,
-        solution_steps=None,
-    )
-
-    return QuestionGenerationResponse(
-        questions=[sample_question],
-        generation_id=str(uuid.uuid4()),
-        generated_at=datetime.now(UTC),
-    )
-
-
-@router.post("/questions/difficulty/control")
-async def control_question_difficulty(
-    request: DifficultyControlRequest,
-) -> DifficultyControlResponse:
-    """
-    Adjust question difficulty based on educational constraints
-    """
-    # Mock implementation
-    return DifficultyControlResponse(
-        adjusted_question=request.question_text,
-        difficulty_score=0.7,
-        cognitive_complexity=0.65,
-        step_count=3,
-    )
-
-
-@router.post("/questions/distractors/generate")
-async def generate_distractors(
-    question_text: str, knowledge_point_id: str
-) -> dict[str, Any]:
-    """
-    Generate cognitively appropriate distractors for multiple choice questions
-    """
-    # Mock implementation
-    _ = question_text, knowledge_point_id  # Avoid unused variable warnings
-    return {
-        "distractors": [
-            {"text": "Common mistake 1", "mistake_pattern": "conceptual_error"},
-            {"text": "Common mistake 2", "mistake_pattern": "calculation_error"},
-        ]
-    }
-
-
-# ============ Assessment & Feedback Endpoints ============
-@router.post("/assessment/evaluate")
-async def evaluate_answers(request: AssessmentRequest) -> AssessmentResponse:
-    """
-    Evaluate student answers and provide detailed assessment
-    """
-    # Mock implementation
-    results = [
-        AssessmentResult(
-            question_id=submission.question_id,
-            is_correct=True,
-            score=1.0,
-            feedback="Good job!",
-            correct_answer="Sample answer",
-            explanation="Sample explanation",
-            mistake_pattern=None,
-        )
-        for submission in request.submissions
-    ]
-
-    return AssessmentResponse(
-        results=results,
-        total_score=len(results),
-        accuracy=1.0,
-        weak_knowledge_points=[],
-        common_mistakes=[],
-    )
-
-
-@router.post("/feedback/provide")
-async def provide_feedback(request: FeedbackRequest) -> FeedbackResponse:
-    """
-    Provide detailed feedback on student answers
-    """
-    # Mock implementation
-    _ = request  # Avoid unused variable warning
-    return FeedbackResponse(
-        score=0.8,
-        detailed_feedback="Good attempt with minor errors",
-        suggestions=["Review concept X", "Practice more problems"],
-        recommended_practice=["Exercise 1", "Exercise 2"],
-    )
-
-
-# ============ User Management Endpoints ============
-@router.post("/users/register")
-async def register_user(request: UserCreateRequest) -> UserResponse:
-    """
-    Register a new user (student/teacher/admin)
-    """
-    # Mock implementation
-    return UserResponse(
-        id=str(uuid.uuid4()),
-        username=request.username,
-        email=request.email,
-        role=request.role,
-        grade_level=request.grade_level,
-        subject_interests=request.subject_interests,
-        created_at=datetime.now(UTC),
-        is_active=True,
-    )
-
-
-@router.post("/users/login")
-async def login_user(request: LoginRequest) -> LoginResponse:
-    """
-    User login and authentication
-    """
-    # Mock implementation
-    user = UserResponse(
-        id=str(uuid.uuid4()),
-        username=request.username,
-        email=f"{request.username}@example.com",
-        role=UserRole.STUDENT,
-        grade_level=None,
-        subject_interests=None,
-        created_at=datetime.now(UTC),
-        is_active=True,
-    )
-
-    return LoginResponse(access_token="sample_token", user=user)
-
-
-@router.get("/users/{user_id}")
-async def get_user_profile(user_id: str) -> UserResponse:
-    """
-    Get user profile information
-    """
-    # Mock implementation
-    return UserResponse(
-        id=user_id,
-        username="sample_user",
-        email="user@example.com",
-        role=UserRole.STUDENT,
-        grade_level=None,
-        subject_interests=None,
-        created_at=datetime.now(UTC),
-        is_active=True,
-    )
-
-
-# ============ Exercise & Practice Endpoints ============
-@router.post("/exercises")
-async def create_exercise(request: ExerciseCreateRequest) -> ExerciseResponse:
-    """
-    Create a new exercise with specific constraints
-    """
-    # Mock implementation
-    return ExerciseResponse(
-        id=str(uuid.uuid4()),
-        title=request.title,
-        description=request.description,
-        subject=request.subject,
-        difficulty=request.difficulty,
-        question_ids=[str(uuid.uuid4()) for _ in range(request.num_questions)],
-        created_at=datetime.now(UTC),
-        created_by="teacher_123",
-    )
-
-
-@router.post("/practice/session")
-async def start_practice_session(
-    request: PracticeSessionRequest,
-) -> PracticeSessionResponse:
-    """
-    Start a new practice session for students
-    """
-    # Mock implementation
-    questions = [
-        GeneratedQuestion(
-            id=str(uuid.uuid4()),
-            question_text=f"Practice question {i + 1}",
-            question_type=QuestionType.MULTIPLE_CHOICE,
-            difficulty=request.difficulty,
-            cognitive_level=CognitiveLevel.UNDERSTANDING,
-            knowledge_point_ids=request.knowledge_point_ids or ["kp_123"],
-            estimated_difficulty=0.6,
-            options=None,
-            correct_answer=None,
-            explanation=None,
-            solution_steps=None,
-        )
-        for i in range(request.num_questions)
-    ]
-
-    return PracticeSessionResponse(
-        session_id=str(uuid.uuid4()),
-        questions=questions,
-        started_at=datetime.now(UTC),
-        time_limit=30,
-    )
-
-
-@router.get("/exercises/{exercise_id}")
-async def get_exercise(exercise_id: str) -> ExerciseResponse:
-    """
-    Get exercise details
-    """
-    # Mock implementation
-    return ExerciseResponse(
-        id=exercise_id,
-        title="Sample Exercise",
-        description="Sample exercise description",
-        subject=SubjectArea.MATH,
-        difficulty=DifficultyLevel.MEDIUM,
-        question_ids=[str(uuid.uuid4()) for _ in range(10)],
-        created_at=datetime.now(UTC),
-        created_by="teacher_123",
-    )
-
-
-# ============ Analytics & Reporting Endpoints ============
-@router.post("/analytics/performance")
-async def get_performance_analytics(
-    request: PerformanceAnalyticsRequest,
-) -> PerformanceAnalyticsResponse:
-    """
-    Get performance analytics for students or classes
-    """
-    # Mock implementation
-    _ = request  # Avoid unused variable warning
-    return PerformanceAnalyticsResponse(
-        overall_accuracy=0.75,
-        total_attempts=100,
-        average_time_per_question=45.5,
-        knowledge_point_analytics=[],
-        progress_trend=[],
-        weak_areas=["algebra", "geometry"],
-    )
-
-
-@router.post("/analytics/mistakes")
-async def analyze_mistakes(request: MistakeAnalysisRequest) -> MistakeAnalysisResponse:
-    """
-    Analyze mistake patterns and provide remediation suggestions
-    """
-    # Mock implementation
-    _ = request  # Avoid unused variable warning
-    return MistakeAnalysisResponse(
-        mistake_patterns=[
-            {"pattern": "conceptual_error", "frequency": 15},
-            {"pattern": "calculation_error", "frequency": 8},
-        ],
-        frequency_distribution={"conceptual_error": 15, "calculation_error": 8},
-        recommended_remediation=[
-            "Review basic concepts",
-            "Practice more calculation problems",
-        ],
-        similar_questions=["q_123", "q_456"],
-    )
-
-
-@router.get("/analytics/class/{class_id}")
-async def get_class_analytics(
-    class_id: str, time_period: str | None = "30d"
-) -> dict[str, Any]:
-    """
-    Get analytics for entire class
-    """
-    # Mock implementation
-    return {
-        "class_id": class_id,
-        "average_accuracy": 0.68,
-        "top_performers": ["student_1", "student_2"],
-        "common_challenges": ["topic_1", "topic_2"],
-        "time_period": time_period,
-    }
-
-
-# ============ System Endpoints ============
-@router.get("/health")
-async def health_check() -> HealthCheckResponse:
-    """
-    System health check endpoint
-    """
-    return HealthCheckResponse(
-        status="healthy",
-        version="1.0.0",
-        timestamp=datetime.now(UTC),
-        components={
-            "database": "connected",
-            "llm_service": "available",
-            "knowledge_graph": "operational",
-        },
-    )
-
-
-@router.post("/batch/questions/generate")
-async def batch_generate_questions(
-    requests: list[QuestionGenerationRequest],
-) -> BatchOperationResponse:
-    """
-    Batch generate questions for multiple knowledge points
-    """
-    # Mock implementation
-    return BatchOperationResponse(
-        operation_id=str(uuid.uuid4()),
-        status="processing",
-        processed_count=len(requests),
-        success_count=len(requests),
-        failed_count=0,
-    )
-
-
-# Error handling
-@router.get("/error")
-async def trigger_error() -> dict[str, str]:
-    """
-    Example error endpoint for testing
-    """
-    raise HTTPException(status_code=400, detail="Sample error for testing")
-
-
-# Include all routers
-api_routers = [router]
-
-# Export for main app
-__all__ = ["api_routers"]
