+++ eduagent/api/endpoints/analytics.py
@@ -1,0 +1,117 @@
+import uuid
+from datetime import UTC, datetime
+from typing import Any
+
+from fastapi import APIRouter, HTTPException
+
+from eduagent.api.schemas import (
+    BatchOperationResponse,
+    HealthCheckResponse,
+    MistakeAnalysisRequest,
+    MistakeAnalysisResponse,
+    PerformanceAnalyticsRequest,
+    PerformanceAnalyticsResponse,
+    QuestionGenerationRequest,
+)
+
+router = APIRouter()
+
+
+@router.post("/analytics/performance")
+async def get_performance_analytics(
+    request: PerformanceAnalyticsRequest,
+) -> PerformanceAnalyticsResponse:
+    """
+    Get performance analytics for students or classes
+    """
+    # Mock implementation
+    _ = request  # Avoid unused variable warning
+    return PerformanceAnalyticsResponse(
+        overall_accuracy=0.75,
+        total_attempts=100,
+        average_time_per_question=45.5,
+        knowledge_point_analytics=[],
+        progress_trend=[],
+        weak_areas=["algebra", "geometry"],
+    )
+
+
+@router.post("/analytics/mistakes")
+async def analyze_mistakes(request: MistakeAnalysisRequest) -> MistakeAnalysisResponse:
+    """
+    Analyze mistake patterns and provide remediation suggestions
+    """
+    # Mock implementation
+    _ = request  # Avoid unused variable warning
+    return MistakeAnalysisResponse(
+        mistake_patterns=[
+            {"pattern": "conceptual_error", "frequency": 15},
+            {"pattern": "calculation_error", "frequency": 8},
+        ],
+        frequency_distribution={"conceptual_error": 15, "calculation_error": 8},
+        recommended_remediation=[
+            "Review basic concepts",
+            "Practice more calculation problems",
+        ],
+        similar_questions=["q_123", "q_456"],
+    )
+
+
+@router.get("/analytics/class/{class_id}")
+async def get_class_analytics(
+    class_id: str, time_period: str | None = "30d"
+) -> dict[str, Any]:
+    """
+    Get analytics for entire class
+    """
+    # Mock implementation
+    return {
+        "class_id": class_id,
+        "average_accuracy": 0.68,
+        "top_performers": ["student_1", "student_2"],
+        "common_challenges": ["topic_1", "topic_2"],
+        "time_period": time_period,
+    }
+
+
+@router.get("/health")
+async def health_check() -> HealthCheckResponse:
+    """
+    System health check endpoint
+    """
+    return HealthCheckResponse(
+        status="healthy",
+        version="1.0.0",
+        timestamp=datetime.now(UTC),
+        components={
+            "database": "connected",
+            "llm_service": "available",
+            "knowledge_graph": "operational",
+        },
+    )
+
+
+@router.post("/batch/questions/generate")
+async def batch_generate_questions(
+    requests: list[QuestionGenerationRequest],
+) -> BatchOperationResponse:
+    """
+    Batch generate questions for multiple knowledge points
+    """
+    # Mock implementation
+    return BatchOperationResponse(
+        operation_id=str(uuid.uuid4()),
+        status="processing",
+        processed_count=len(requests),
+        success_count=len(requests),
+        failed_count=0,
+    )
+
+
+# Error handling
+@router.get("/error")
+async def trigger_error() -> dict[str, str]:
+    """
+    Example error endpoint for testing
+    """
+    raise HTTPException(status_code=400, detail="Sample error for testing")
