+++ eduagent/api/endpoints/assessment.py
@@ -1,0 +1,54 @@
+from fastapi import APIRouter
+
+from eduagent.api.schemas import (
+    AssessmentRequest,
+    AssessmentResponse,
+    AssessmentResult,
+    FeedbackRequest,
+    FeedbackResponse,
+)
+
+router = APIRouter()
+
+
+@router.post("/assessment/evaluate")
+async def evaluate_answers(request: AssessmentRequest) -> AssessmentResponse:
+    """
+    Evaluate student answers and provide detailed assessment
+    """
+    # Mock implementation
+    results = [
+        AssessmentResult(
+            question_id=submission.question_id,
+            is_correct=True,
+            score=1.0,
+            feedback="Good job!",
+            correct_answer="Sample answer",
+            explanation="Sample explanation",
+            mistake_pattern=None,
+        )
+        for submission in request.submissions
+    ]
+
+    return AssessmentResponse(
+        results=results,
+        total_score=len(results),
+        accuracy=1.0,
+        weak_knowledge_points=[],
+        common_mistakes=[],
+    )
+
+
+@router.post("/feedback/provide")
+async def provide_feedback(request: FeedbackRequest) -> FeedbackResponse:
+    """
+    Provide detailed feedback on student answers
+    """
+    # Mock implementation
+    _ = request  # Avoid unused variable warning
+    return FeedbackResponse(
+        score=0.8,
+        detailed_feedback="Good attempt with minor errors",
+        suggestions=["Review concept X", "Practice more problems"],
+        recommended_practice=["Exercise 1", "Exercise 2"],
+    )
