+++ eduagent/api/endpoints/knowledge.py
@@ -1,0 +1,53 @@
+import uuid
+from datetime import UTC, datetime
+
+from fastapi import APIRouter
+
+from eduagent.api.schemas import (
+    KnowledgeExtractionResponse,
+    KnowledgeGraphResponse,
+)
+
+router = APIRouter()
+
+
+@router.post("/textbook/upload")
+async def upload_textbook() -> KnowledgeExtractionResponse:
+    """
+    Upload textbook for knowledge extraction and multi-modal analysis
+    """
+    extraction_id = str(uuid.uuid4())
+
+    # Background task for processing would be added here
+    return KnowledgeExtractionResponse(
+        extraction_id=extraction_id,
+        status="processing",
+        extracted_concepts=[],
+        created_at=datetime.now(UTC),
+    )
+
+
+@router.get("/knowledge/extraction/{extraction_id}")
+async def get_extraction_status(extraction_id: str) -> KnowledgeExtractionResponse:
+    """
+    Get status of knowledge extraction process
+    """
+    # Mock implementation
+    return KnowledgeExtractionResponse(
+        extraction_id=extraction_id,
+        status="completed",
+        extracted_concepts=[{"concept": "sample", "confidence": 0.95}],
+        created_at=datetime.now(UTC),
+    )
+
+
+@router.get("/knowledge/graph/{textbook_id}")
+async def get_knowledge_graph(textbook_id: str) -> KnowledgeGraphResponse:
+    """
+    Retrieve 3D knowledge graph for a textbook
+    """
+    # Mock implementation
+    _ = textbook_id  # Avoid unused variable warning
+    return KnowledgeGraphResponse(
+        knowledge_points=[], ability_targets=[], common_mistakes=[], graph_structure={}
+    )
