+++ eduagent/api/endpoints/questions.py
@@ -1,0 +1,78 @@
+import uuid
+from datetime import UTC, datetime
+from typing import Any
+
+from fastapi import APIRouter
+
+from eduagent.api.schemas import (
+    CognitiveLevel,
+    DifficultyControlRequest,
+    DifficultyControlResponse,
+    GeneratedQuestion,
+    QuestionGenerationRequest,
+    QuestionGenerationResponse,
+)
+
+router = APIRouter()
+
+
+@router.post("/questions/generate")
+async def generate_questions(
+    request: QuestionGenerationRequest,
+) -> QuestionGenerationResponse:
+    """
+    Generate educational questions based on knowledge points and constraints
+    """
+    # Mock implementation
+    sample_question = GeneratedQuestion(
+        id=str(uuid.uuid4()),
+        question_text="Sample generated question",
+        question_type=request.question_type,
+        difficulty=request.difficulty,
+        cognitive_level=CognitiveLevel.UNDERSTANDING,
+        knowledge_point_ids=request.knowledge_point_ids,
+        estimated_difficulty=0.6,
+        options=None,
+        correct_answer=None,
+        explanation=None,
+        solution_steps=None,
+    )
+
+    return QuestionGenerationResponse(
+        questions=[sample_question],
+        generation_id=str(uuid.uuid4()),
+        generated_at=datetime.now(UTC),
+    )
+
+
+@router.post("/questions/difficulty/control")
+async def control_question_difficulty(
+    request: DifficultyControlRequest,
+) -> DifficultyControlResponse:
+    """
+    Adjust question difficulty based on educational constraints
+    """
+    # Mock implementation
+    return DifficultyControlResponse(
+        adjusted_question=request.question_text,
+        difficulty_score=0.7,
+        cognitive_complexity=0.65,
+        step_count=3,
+    )
+
+
+@router.post("/questions/distractors/generate")
+async def generate_distractors(
+    question_text: str, knowledge_point_id: str
+) -> dict[str, Any]:
+    """
+    Generate cognitively appropriate distractors for multiple choice questions
+    """
+    # Mock implementation
+    _ = question_text, knowledge_point_id  # Avoid unused variable warnings
+    return {
+        "distractors": [
+            {"text": "Common mistake 1", "mistake_pattern": "conceptual_error"},
+            {"text": "Common mistake 2", "mistake_pattern": "calculation_error"},
+        ]
+    }
