+++ eduagent/api/endpoints/users.py
@@ -1,0 +1,11 @@
+# TODO(YunX): define the user management endpoints and pydantic models
+
+"""
+User Management Endpoints
+
+This module provides role-based authentication and user management endpoints:
+- Separate login endpoints for students, teachers, and admins
+- User registration and profile management
+- Session management and token refresh
+- Role-based authorization
+"""
