+++ eduagent/tools/__init__.py
@@ -1,8 +1,1 @@
-# Tool interfaces for educational AI system
-from .analytics_tool import AnalyticsTool
-from .assessment_tool import AssessmentTool
-from .base import BaseTool
-from .question_tool import QuestionGenerationTool
-from .rag_tool import RAGTool
-
-__all__ = ["AnalyticsTool", "AssessmentTool", "BaseTool", "QuestionGenerationTool", "RAGTool"]
+# LLM Tools for educational AI system
