+++ eduagent/user/__init__.py
@@ -1,0 +1,12 @@
+# TODO(YunX): impl user management core logic, see the comment below
+# TODO(Marisolebxf): mpl user login page
+
+"""
+User Service Module
+
+This module handles all user-related business logic including:
+- User authentication and authorization
+- User CRUD operations, create a class like UserDatabaseService, to do the CRUD operations
+- Role-based access control
+- Session management
+"""
